feat: refactor plugin name to samsung_galaxy_in_app_purchase
