+++ android/src/main/AndroidManifest.xml
@@ -1,4 +1,4 @@
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
-  package="me.genopets.plugins.galaxy_iap">
-  <uses-permission android:name="com.samsung.android.iap.permission.BILLING"/>
+  package="me.genopets.plugins.samsung_galaxy_in_app_purchase">
+  <uses-permission android:name="com.samsung.android.iap.permission.BILLING" />
 </manifest>
