+++ example/lib/dialogs/consumes_list_dialog.dart
@@ -1,5 +1,5 @@
 import 'package:flutter/material.dart';
-import 'package:galaxy_iap/entities/galaxy_consume.dart';
+import 'package:samsung_galaxy_in_app_purchase/entities/galaxy_consume.dart';
 
 class ConsumesListDialog extends StatelessWidget {
   final List<GalaxyConsume> consumes;
