+++ example/pubspec.yaml
@@ -17,9 +17,9 @@ dependencies:
   flutter:
     sdk: flutter
 
-  galaxy_iap:
+  samsung_galaxy_in_app_purchase:
     # When depending on this package from a real application you should use:
-    #   galaxy_iap: ^x.y.z
+    #   samsung_galaxy_in_app_purchase: ^x.y.z
     # See https://dart.dev/tools/pub/dependencies#version-constraints
     # The example app is bundled with the plugin so we use a path dependency on
     # the parent directory to use the current plugin's version.
