+++ lib/galaxy_iap_platform_interface.dart
@@ -1,8 +1,8 @@
-import 'package:galaxy_iap/entities/galaxy_consume.dart';
-import 'package:galaxy_iap/entities/galaxy_owned_product.dart';
-import 'package:galaxy_iap/entities/galaxy_product.dart';
-import 'package:galaxy_iap/entities/galaxy_purchase.dart';
-import 'package:galaxy_iap/galaxy_iap.dart';
+import 'package:samsung_galaxy_in_app_purchase/entities/galaxy_consume.dart';
+import 'package:samsung_galaxy_in_app_purchase/entities/galaxy_owned_product.dart';
+import 'package:samsung_galaxy_in_app_purchase/entities/galaxy_product.dart';
+import 'package:samsung_galaxy_in_app_purchase/entities/galaxy_purchase.dart';
+import 'package:samsung_galaxy_in_app_purchase/samsung_galaxy_in_app_purchase.dart';
 import 'package:plugin_platform_interface/plugin_platform_interface.dart';
 
 import 'galaxy_iap_method_channel.dart';
