+++ pubspec.yaml
@@ -1,4 +1,4 @@
-name: galaxy_iap
+name: samsung_galaxy_in_app_purchase
 description: A new Flutter plugin project.
 version: 0.0.1
 homepage:
@@ -35,7 +35,7 @@ flutter:
   plugin:
     platforms:
       android:
-        package: me.genopets.plugins.galaxy_iap
+        package: me.genopets.plugins.samsung_galaxy_in_app_purchase
         pluginClass: GalaxyIapPlugin
 
   # To add assets to your plugin package, add an assets section, like this:
